+++ src/fields/Content.php
@@ -4,8 +4,10 @@ namespace verbb\formie\fields;
 use verbb\formie\base\CosmeticField;
 use verbb\formie\elements\Form;
 use verbb\formie\elements\Submission;
+use verbb\formie\helpers\Gql as FormieGql;
 use verbb\formie\helpers\RichTextHelper;
 use verbb\formie\helpers\SchemaHelper;
+use verbb\formie\gql\types\Json as JsonType;
 use verbb\formie\models\Notification;
 use verbb\formie\models\RichText;
 use verbb\formie\models\SlotTag;
@@ -143,6 +145,11 @@ class Content extends CosmeticField
                 'resolve' => function(self $field) {
                     return $field->content->toHtml(null, false);
                 },
+            ],
+            'contentJson' => [
+                'name' => 'contentJson',
+                'type' => JsonType::getType(),
+                'resolve' => static fn(self $field) => FormieGql::resolveRichTextJson($field->content),
             ],
         ]);
     }